Rename pale_fire::color to pale_fire::style
